add users to work schedule

+++ app/Http/Controllers/TimeRecordingSystems/TimeRecordingSystemScheduleWorkScheduleAssignmentUserController.php
@@ -8,6 +8,7 @@ use App\Models\WorkSchedule;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Models\WorkScheduleAssignment;
+use Illuminate\Support\Facades\Validator;
 use App\Helpers\AddDefaultWorkScheduleAssignment;
 use App\Services\UpdatedRoleGroupCollectionService;
 
@@ -24,35 +25,37 @@ class TimeRecordingSystemScheduleWorkScheduleAssignmentUserController extends Co
     public function index($scheduleId,$year,$monthId)
     {
         $action = 'show';
-        $groupUrl = session('groupUrl');
+        $groupUrl = strval(session('groupUrl'));
         $roleGroupCollection = $this->updatedRoleGroupCollectionService->getUpdatedRoleGroupCollection($action);
         $updatedRoleGroupCollection = $roleGroupCollection['updatedRoleGroupCollection'];
         $permission = $roleGroupCollection['permission'];
         $workSchedule = WorkSchedule::find($scheduleId);
         $users = $this->getUsersByWorkScheduleAssignment($scheduleId, $monthId, $year)->paginate(20);
-        // dd($this->getUsersByWorkScheduleAssignment($scheduleId, $monthId, $year)->get());
+
         return view('groups.time-recording-system.schedulework.schedule.assignment.user.index', [
             'groupUrl' => $groupUrl,
             'modules' => $updatedRoleGroupCollection,
+            'permission' => $permission,
             'users' => $users,
             'workSchedule' => $workSchedule,
             'year' => $year,
-            'monthId' => $monthId,
-            'permission' => $permission
+            'monthId' => $monthId
         ]);
     }
 
     public function create($scheduleId,$year,$monthId)
     {
         $action = 'create';
-        $groupUrl = session('groupUrl');
+        $groupUrl = strval(session('groupUrl'));
         $roleGroupCollection = $this->updatedRoleGroupCollectionService->getUpdatedRoleGroupCollection($action);
         $updatedRoleGroupCollection = $roleGroupCollection['updatedRoleGroupCollection'];
+        $permission = $roleGroupCollection['permission'];
         $workSchedule = WorkSchedule::find($scheduleId);
         $users = User::paginate(20);
         return view('groups.time-recording-system.schedulework.schedule.assignment.user.create', [
             'groupUrl' => $groupUrl,
             'modules' => $updatedRoleGroupCollection,
+            'permission' => $permission,
             'users' => $users,
             'workSchedule' => $workSchedule,
             'year' => $year,
@@ -61,25 +64,47 @@ class TimeRecordingSystemScheduleWorkScheduleAssignmentUserController extends Co
     }
     public function store(Request $request)
     {
-        $workScheduleId = $request->workScheduleId;
-        $year = $request->year;
-        $month = $request->month;
-        
-        $selectedUsers = $request->users;
-        $users = User::whereIn('id',$selectedUsers)->get();
+        $validator = Validator::make($request->all(), [
+            'workScheduleId' => 'required',
+            'year' => 'required',
+            'month' => 'required',
+            'users' => 'required|array',
+            'users.*' => 'exists:users,id',
+        ], [
+            'workScheduleId.required' => 'กรุณากรอกข้อมูล Work Schedule ID',
+            'year.required' => 'กรุณากรอกข้อมูลปี',
+            'month.required' => 'กรุณากรอกข้อมูลเดือน',
+            'users.required' => 'กรุณาเลือกพนักงาน',
+            'users.array' => 'ข้อมูลผู้ใช้ต้องเป็นรูปแบบของอาร์เรย์',
+            'users.*.exists' => 'ผู้ใช้บางรายที่เลือกไม่มีอยู่ในระบบ',
+        ]);
 
-        foreach($users as $user)
-        {
-            $workScheduleAssignments = WorkScheduleAssignment::where('work_schedule_id',$workScheduleId)->where('month_id',$month)->where('year',$year)->get();
-            // dd($workScheduleAssignments);
-            $user->workScheduleAssignments()->detach(); 
-            $user->workScheduleAssignments()->attach($workScheduleAssignments); 
+        if ($validator->fails()) {
+            // Handle validation failure, return error response, or perform any necessary actions
+            return redirect()->back()->withErrors($validator)->withInput();
+            // ...
+        } else {
+            $workScheduleId = $request->workScheduleId;
+            $year = $request->year;
+            $month = $request->month;
+            
+            $selectedUsers = $request->users;
+        
+            $users = User::whereIn('id',$selectedUsers)->get();
+
+            foreach($users as $user)
+            {
+                $workScheduleAssignments = WorkScheduleAssignment::where('work_schedule_id',$workScheduleId)->where('month_id',$month)->where('year',$year)->get();
+                // dd($workScheduleAssignments);
+                $user->workScheduleAssignments()->detach(); 
+                $user->workScheduleAssignments()->attach($workScheduleAssignments); 
+            }
+            
+            $url = "groups/time-recording-system/schedulework/schedule/assignment/user/{$workScheduleId}/year/{$year}/month/{$month}";
+            return redirect()->to($url);
         }
-        
-        $url = "groups/time-recording-system/schedulework/schedule/assignment/user/{$workScheduleId}/year/{$year}/month/{$month}";
-        return redirect()->to($url);
-
     }
+
     /**
      * ค้นหาพนักงงาน
      *
@@ -113,17 +138,6 @@ class TimeRecordingSystemScheduleWorkScheduleAssignmentUserController extends Co
 
     public function delete($workScheduleId,$year,$monthId, $userId)
     {
-        
-        // $workScheduleAssignment = WorkScheduleAssignment::where('work_schedule_id', $workScheduleId)
-        //     ->where('month_id', $month)
-        //     ->where('year', $year)
-        //     ->first();
-
-        // if ($workScheduleAssignment) {
-        //         $workScheduleAssignment->workScheduleAssignmentUsers()
-        //             ->where('user_id', $userId)
-        //             ->sync([]);
-        //     }
         $user = User::find($userId);
         if ($user) {
             $user->workScheduleAssignments()
@@ -132,8 +146,6 @@ class TimeRecordingSystemScheduleWorkScheduleAssignmentUserController extends Co
                 ->where('year', $year)
                 ->detach();
         }
-
-
 
         $url = "groups/time-recording-system/schedulework/schedule/assignment/user/{$workScheduleId}/year/{$year}/month/{$monthId}";
         return redirect()->to($url);
